requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
matplotlib=3.2.1=0
numpy=1.19.4
opencv=4.2.0
pandas=1.1.1
pillow=7.1.2
python=3.8.2
pytorch=1.7.0
scikit-image=0.17.2
scikit-learn=0.22.1
scipy=1.4.1
tensorboard=2.2.1
torchvision=0.8.1
tqdm=4.45.0
fvcore=0.1.5.post20220512
